+++ client/src/routes/JoinSessionModal/JoinSessionModal.tsx
@@ -8,7 +8,7 @@ import {JoinSessionError} from '../../../../shared/src/errors/Session';
 import {COLORS} from '../../../../shared/src/constants/colors';
 
 import Gutters from '../../common/components/Gutters/Gutters';
-import HalfModal from '../../common/components/Modals/HalfModal';
+import Modal from '../../common/components/Modal/Modal';
 import {Spacer16, Spacer24} from '../../common/components/Spacers/Spacer';
 import {Body14} from '../../common/components/Typography/Body/Body';
 import {Display24} from '../../common/components/Typography/Display/Display';
@@ -32,7 +32,7 @@ const JoinSessionModal = () => {
   const [errorString, setErrorString] = useState<string | null>(null);
 
   return (
-    <HalfModal>
+    <Modal>
       <Spacer16 />
       <Gutters>
         <Display24>{t('title')}</Display24>
@@ -65,7 +65,7 @@ const JoinSessionModal = () => {
           </>
         )}
       </Gutters>
-    </HalfModal>
+    </Modal>
   );
 };
