+++ client/src/routes/SessionModal/SessionModal.tsx
@@ -15,7 +15,7 @@ import {
   ShareIcon,
 } from '../../common/components/Icons';
 import Image from '../../common/components/Image/Image';
-import HalfModal from '../../common/components/Modals/HalfModal';
+import Modal from '../../common/components/Modal/Modal';
 import {Spacer16, Spacer8} from '../../common/components/Spacers/Spacer';
 import {Display24} from '../../common/components/Typography/Display/Display';
 import {COLORS} from '../../../../shared/src/constants/colors';
@@ -119,7 +119,7 @@ const SessionModal = () => {
   };
 
   return (
-    <HalfModal>
+    <Modal>
       <Spacer16 />
       <Content>
         <View>
@@ -188,7 +188,7 @@ const SessionModal = () => {
           <DeleteButton small onPress={onDelete} Icon={DeleteIcon} />
         )}
       </BottomContent>
-    </HalfModal>
+    </Modal>
   );
 };
