+++ client/src/routes/Sessions/components/AddSessionModal.tsx
@@ -3,26 +3,27 @@ import {useNavigation} from '@react-navigation/native';
 import {NativeStackNavigationProp} from '@react-navigation/native-stack';
 import {useTranslation} from 'react-i18next';
 
-import Gutters from '../../common/components/Gutters/Gutters';
-import Modal from '../../common/components/Modal/Modal';
+import Gutters from '../../../common/components/Gutters/Gutters';
+import Modal from '../../../common/components/Modal/Modal';
 import {
   Spacer16,
   Spacer24,
   Spacer8,
-} from '../../common/components/Spacers/Spacer';
-import {Display24} from '../../common/components/Typography/Display/Display';
+} from '../../../common/components/Spacers/Spacer';
+import {Display24} from '../../../common/components/Typography/Display/Display';
 
-import {ModalStackProps} from '../../lib/navigation/constants/routes';
-import Button from '../../common/components/Buttons/Button';
+import {ModalStackProps} from '../../../lib/navigation/constants/routes';
+import Button from '../../../common/components/Buttons/Button';
 
 const AddSessionModal: React.FC = () => {
-  const {t} = useTranslation('Modal.AddSession');
+  const {t} = useTranslation('Component.AddSessionModal');
 
   const {popToTop, navigate} =
     useNavigation<NativeStackNavigationProp<ModalStackProps>>();
 
   return (
     <Modal>
+      <Spacer16 />
       <Gutters>
         <Display24>{t('title')}</Display24>
         <Spacer24 />
